Add files via upload
India is a big country where the migration rate is very high. Due to this many of the migrants does not have a clear idea about the medical services around them. Also many people are unaware of the government schemes for availing medical facilities. MediGuide web application helps those people in need who are unaware of the medical services around them. This application helps people to check for the availability of medical facilities, based on search criteria and related government schemes .